+++ framework/contrib/hit/Makefile
@@ -18,9 +18,13 @@ endif
 
 $(info Building hit for python with $(pyconfig))
 
+hit_srcdir         := src/hit
+hit_srcfiles       := $(hit_srcdir)/parse.cc $(hit_srcdir)/lex.cc $(hit_srcdir)/braceexpr.cc
+hit_includeflags   := -Iinclude
+
 PYTHONPREFIX ?= `$(pyconfig) --prefix`
 PYTHONCFLAGS ?= `$(pyconfig) --cflags`
-HITCPP := hit.cpp
+HITCPP := $(hit_srcdir)/hit.cpp
 
 WASP_DIR           ?= $(abspath ../wasp/install)
 
@@ -42,13 +46,13 @@ endif
 wasp_CXXFLAGS  += -I$(WASP_DIR)/include
 wasp_LDFLAGS   += -Wl,-rpath,$(WASP_DIR)/lib -L$(WASP_DIR)/lib $(wasp_LIBS)
 
-hit: main.cc parse.cc lex.cc braceexpr.cc braceexpr.h lex.h parse.h
-	$(CXX) -std=c++17 $(wasp_CXXFLAGS) -g $(CXXFLAGS) $< parse.cc lex.cc braceexpr.cc -o $@ $(wasp_LDFLAGS)
+hit: $(hit_srcdir)/main.cc $(hit_srcfiles) include/hit/braceexpr.h include/hit/lex.h include/hit/parse.h
+	$(CXX) -std=c++17 $(wasp_CXXFLAGS) $(hit_includeflags) -g $(CXXFLAGS) $< $(hit_srcfiles) -o $@ $(wasp_LDFLAGS)
 
 bindings: hit.so
 
-hit.so: parse.cc lex.cc braceexpr.cc
-	$(CXX) -std=c++17 $(wasp_CXXFLAGS) -w -fPIC -lstdc++ -shared -L$(PYTHONPREFIX)/lib $(PYTHONCFLAGS) $(DYNAMIC_LOOKUP) $^ $(HITCPP) -o $@ $(wasp_LDFLAGS)
+hit.so: $(hit_srcfiles)
+	$(CXX) -std=c++17 $(wasp_CXXFLAGS) $(hit_includeflags) -w -fPIC -lstdc++ -shared -L$(PYTHONPREFIX)/lib $(PYTHONCFLAGS) $(DYNAMIC_LOOKUP) $^ $(HITCPP) -o $@ $(wasp_LDFLAGS)
 
 $(HITCPP): hit.pyx chit.pxd
 	$(cython) -o $@ --cplus $<
