+++ framework/contrib/hit/include/hit/hit.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "parse.h"
-#include "braceexpr.h"
+#include "hit/parse.h"
+#include "hit/braceexpr.h"
 
 #include <fstream> // TODO: remove this - hit parser doesn't need it at all
