+++ framework/contrib/hit/src/hit/braceexpr.cc
@@ -1,4 +1,4 @@
-#include "braceexpr.h"
+#include "hit/braceexpr.h"
 
 #include <cstdlib>
 #include <iostream>
