+++ unit/src/HitTests.C
@@ -7,7 +7,7 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "hit.h"
+#include "hit/hit.h"
 #include "Parser.h"
 
 #include "gtest_include.h"
